GeminiAgent/requirements.txt
google-generativeai
pexpect
pillow
rich
duckduckgo-search
requests
